requirements.txt
allure-pytest
pytest
pytest-testconfig
requests
urllib3